+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/ApiDuplicatorServiceImpl.java
@@ -23,10 +23,8 @@ import static java.util.stream.Collectors.*;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.gravitee.common.http.HttpMethod;
-import io.gravitee.definition.model.DefinitionContext;
 import io.gravitee.definition.model.DefinitionVersion;
 import io.gravitee.definition.model.Proxy;
 import io.gravitee.definition.model.VirtualHost;
@@ -45,9 +43,9 @@ import io.gravitee.rest.api.service.exceptions.*;
 import io.gravitee.rest.api.service.imports.ImportApiJsonNode;
 import io.gravitee.rest.api.service.imports.ImportJsonNode;
 import io.gravitee.rest.api.service.imports.ImportJsonNodeWithIds;
-import io.gravitee.rest.api.service.imports.ImportPlanJsonNode;
 import io.gravitee.rest.api.service.sanitizer.UrlSanitizerUtils;
 import io.gravitee.rest.api.service.spring.ImportConfiguration;
+import io.gravitee.rest.api.service.ApiIdsCalculatorService;
 import io.vertx.core.buffer.Buffer;
 import java.io.IOException;
 import java.util.*;
@@ -88,6 +86,7 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
     private final ApiConverter apiConverter;
     private final PlanConverter planConverter;
     private final PermissionService permissionService;
+    private final ApiIdsCalculatorService apiIdsCalculatorService;
 
     public ApiDuplicatorServiceImpl(
         HttpClientService httpClientService,
@@ -104,7 +103,8 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
         @Lazy ApiService apiService,
         ApiConverter apiConverter,
         PlanConverter planConverter,
-        PermissionService permissionService
+        PermissionService permissionService,
+        ApiIdsCalculatorService apiIdsCalculatorService
     ) {
         this.httpClientService = httpClientService;
         this.importConfiguration = importConfiguration;
@@ -121,6 +121,7 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
         this.apiConverter = apiConverter;
         this.planConverter = planConverter;
         this.permissionService = permissionService;
+        this.apiIdsCalculatorService = apiIdsCalculatorService;
     }
 
     @Override
@@ -128,7 +129,7 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
         String apiDefinition = fetchApiDefinitionContentFromURL(apiDefinitionOrURL);
         try {
             // Read the whole input API definition, and recalculate its ID
-            ImportApiJsonNode apiJsonNode = recalculateApiDefinitionIds(
+            ImportApiJsonNode apiJsonNode = apiIdsCalculatorService.recalculateApiDefinitionIds(
                 executionContext,
                 new ImportApiJsonNode(objectMapper.readTree(apiDefinition))
             );
@@ -159,7 +160,7 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
 
         try {
             // Read the whole input API definition, and recalculate its ID
-            ImportApiJsonNode apiJsonNode = recalculateApiDefinitionIds(
+            ImportApiJsonNode apiJsonNode = apiIdsCalculatorService.recalculateApiDefinitionIds(
                 executionContext,
                 new ImportApiJsonNode(objectMapper.readTree(apiDefinition)),
                 urlApiId
@@ -746,195 +747,6 @@ public class ApiDuplicatorServiceImpl extends AbstractService implements ApiDupl
                 throw new ApiImportException("Only one system folder is allowed in the API pages definition");
             }
         }
-    }
-
-    protected ImportApiJsonNode recalculateApiDefinitionIds(ExecutionContext executionContext, ImportApiJsonNode apiJsonNode) {
-        return recalculateApiDefinitionIds(executionContext, apiJsonNode, null);
-    }
-
-    /*
-     * If the API definition is imported from another environment we need to match the targeted
-     * entities (API, plans, pages) from the input IDs.
-     *
-     * If the API definition holds a cross ID, the entities will be matched using this unique ID.
-     *
-     * If the API definition does not hold a cross ID, the matching entity IDs will be matched using
-     * a predictable ID generation based on the target environment ID, the source API ID and the source entity ID
-     */
-    protected ImportApiJsonNode recalculateApiDefinitionIds(
-        ExecutionContext executionContext,
-        ImportApiJsonNode apiJsonNode,
-        String urlApiId
-    ) {
-        /*
-         * In case of an update, if the API definition ID is the same as the resource ID targeted by the update,
-         * we don't apply any kind of ID transformation so that we don't break previous exports that don't hold
-         * a cross ID
-         */
-        if (!apiJsonNode.hasId() || !apiJsonNode.getId().equals(urlApiId)) {
-            findApiByEnvironmentAndCrossId(executionContext.getEnvironmentId(), apiJsonNode.getCrossId())
-                .ifPresentOrElse(
-                    api -> recalculateIdsFromCrossId(executionContext, apiJsonNode, api),
-                    () -> recalculateIdsFromDefinitionIds(executionContext.getEnvironmentId(), apiJsonNode, urlApiId)
-                );
-        }
-        return generateEmptyIds(apiJsonNode);
-    }
-
-    private Optional<ApiEntity> findApiByEnvironmentAndCrossId(String environmentId, String crossId) {
-        return crossId == null ? Optional.empty() : apiService.findByEnvironmentIdAndCrossId(environmentId, crossId);
-    }
-
-    private void recalculateIdsFromCrossId(ExecutionContext executionContext, ImportApiJsonNode apiJsonNode, ApiEntity api) {
-        apiJsonNode.setId(api.getId());
-        Map<String, String> pagesIdsMap = recalculatePageIdsFromCrossIds(executionContext.getEnvironmentId(), api, apiJsonNode.getPages());
-        recalculatePlanIdsFromCrossIds(executionContext, api, apiJsonNode.getPlans(), pagesIdsMap);
-    }
-
-    private void recalculatePlanIdsFromCrossIds(
-        ExecutionContext executionContext,
-        ApiEntity api,
-        List<ImportPlanJsonNode> plansNodes,
-        Map<String, String> pagesIdsMap
-    ) {
-        Map<String, PlanEntity> plansByCrossId = planService
-            .findByApi(executionContext, api.getId())
-            .stream()
-            .filter(plan -> plan.getCrossId() != null)
-            .collect(toMap(PlanEntity::getCrossId, Function.identity()));
-
-        plansNodes
-            .stream()
-            .filter(ImportJsonNodeWithIds::hasCrossId)
-            .forEach(plan -> {
-                PlanEntity matchingPlan = plansByCrossId.get(plan.getCrossId());
-                plan.setApi(api.getId());
-                plan.setId(matchingPlan != null ? matchingPlan.getId() : UuidString.generateRandom());
-                recalculateGeneralConditionsPageId(plan, pagesIdsMap);
-            });
-    }
-
-    /**
-     * Recalculate imported API pages ID, using crossID.
-     *
-     * @param environmentId
-     * @param api
-     * @param pagesNodes
-     * @return the map of old ID - new ID
-     */
-    private Map<String, String> recalculatePageIdsFromCrossIds(
-        String environmentId,
-        ApiEntity api,
-        List<ImportJsonNodeWithIds> pagesNodes
-    ) {
-        Map<String, String> idsMap = new HashMap<>();
-
-        Map<String, PageEntity> pagesByCrossId = pageService
-            .findByApi(environmentId, api.getId())
-            .stream()
-            .filter(page -> page.getCrossId() != null)
-            .collect(toMap(PageEntity::getCrossId, Function.identity()));
-
-        pagesNodes
-            .stream()
-            .filter(ImportJsonNodeWithIds::hasCrossId)
-            .forEach(page -> {
-                String pageId = page.hasId() ? page.getId() : null;
-                PageEntity matchingPage = pagesByCrossId.get(page.getCrossId());
-                page.setApi(api.getId());
-                if (matchingPage != null) {
-                    idsMap.put(pageId, matchingPage.getId());
-                    page.setId(matchingPage.getId());
-                    updatePagesHierarchy(pagesNodes, pageId, matchingPage.getId());
-                } else {
-                    String newPageId = UuidString.generateRandom();
-                    idsMap.put(pageId, newPageId);
-                    page.setId(newPageId);
-                    updatePagesHierarchy(pagesNodes, pageId, newPageId);
-                }
-            });
-
-        return idsMap;
-    }
-
-    private void recalculateIdsFromDefinitionIds(String environmentId, ImportApiJsonNode apiJsonNode, String urlApiId) {
-        if (canRegenerateId(apiJsonNode)) {
-            String targetApiId = urlApiId == null ? UuidString.generateForEnvironment(environmentId, apiJsonNode.getId()) : urlApiId;
-            apiJsonNode.setId(targetApiId);
-            Map<String, String> pagesIdsMap = recalculatePageIdsFromDefinitionIds(apiJsonNode.getPages(), environmentId, targetApiId);
-            recalculatePlanIdsFromDefinitionIds(apiJsonNode.getPlans(), environmentId, targetApiId, pagesIdsMap);
-        }
-    }
-
-    private void recalculatePlanIdsFromDefinitionIds(
-        List<ImportPlanJsonNode> plansNodes,
-        String environmentId,
-        String apiId,
-        Map<String, String> pagesIdsMap
-    ) {
-        plansNodes
-            .stream()
-            .filter(ImportJsonNodeWithIds::hasId)
-            .forEach(plan -> {
-                plan.setId(UuidString.generateForEnvironment(environmentId, apiId, plan.getId()));
-                plan.setApi(apiId);
-                recalculateGeneralConditionsPageId(plan, pagesIdsMap);
-            });
-    }
-
-    private static void recalculateGeneralConditionsPageId(ImportPlanJsonNode plan, Map<String, String> pagesIdsMap) {
-        if (plan.hasGeneralConditions()) {
-            plan.setGeneralConditions(pagesIdsMap.get(plan.getGeneralConditions()));
-        }
-    }
-
-    /**
-     * Recalculate imported API pages ID, using definition ID.
-     *
-     * @param pagesNodes
-     * @param environmentId
-     * @param apiId
-     * @return the map of old ID - new ID
-     */
-    private Map<String, String> recalculatePageIdsFromDefinitionIds(
-        List<ImportJsonNodeWithIds> pagesNodes,
-        String environmentId,
-        String apiId
-    ) {
-        Map<String, String> idsMap = new HashMap<>();
-        pagesNodes
-            .stream()
-            .filter(ImportJsonNodeWithIds::hasId)
-            .forEach(page -> {
-                String oldPageId = page.getId();
-                String newPageId = UuidString.generateForEnvironment(environmentId, apiId, oldPageId);
-                idsMap.put(oldPageId, newPageId);
-                page.setId(newPageId);
-                page.setApi(apiId);
-                updatePagesHierarchy(pagesNodes, oldPageId, newPageId);
-            });
-        return idsMap;
-    }
-
-    private boolean canRegenerateId(ImportApiJsonNode apiJsonNode) {
-        // If the definition is managed by kubernetes, do not try to recalculate ids because k8s is the source of truth.
-        return !DefinitionContext.ORIGIN_KUBERNETES.equalsIgnoreCase(apiJsonNode.getDefinitionContextOrigin());
-    }
-
-    private void updatePagesHierarchy(List<ImportJsonNodeWithIds> pagesNodes, String parentId, String newParentId) {
-        pagesNodes.stream().filter(child -> isChildPageOf(child, parentId)).forEach(child -> child.setParentId(newParentId));
-    }
-
-    private boolean isChildPageOf(ImportJsonNodeWithIds pageNode, String parentPageId) {
-        return pageNode.hasParentId() && pageNode.getParentId().equals(parentPageId);
-    }
-
-    private ImportApiJsonNode generateEmptyIds(ImportApiJsonNode apiJsonNode) {
-        Stream
-            .concat(apiJsonNode.getPlans().stream(), apiJsonNode.getPages().stream())
-            .filter(node -> !node.hasId())
-            .forEach(node -> node.setId(UuidString.generateRandom()));
-        return apiJsonNode;
     }
 
     private static String stringifyApiDefinitionFromJackson(Object apiDefinitionOrUrl) {
